clean code and add button
