+++ lib/main.dart
@@ -30,12 +30,10 @@ import 'package:flutter/foundation.dart';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:window_manager/window_manager.dart';
 
-import 'package:msfatigue/widgets/page/credentials_page.dart';
 import 'package:msfatigue/features/bloc/survey_bloc.dart';
-import 'package:msfatigue/welcome.dart';
+import 'package:msfatigue/widgets/page/check_credentials.dart';
 
 // Dummy implementations for desktop support.
 // Replace these with your actual implementations.
@@ -71,6 +69,7 @@ Future<void> main() async {
 
     const windowOptions = WindowOptions(
       // Setting alwaysOnTop to true so the app starts on top.
+
       alwaysOnTop: true,
       title: 'MS Fatigue',
     );
@@ -96,6 +95,7 @@ class MSFatigue extends StatelessWidget {
   const MSFatigue({super.key});
 
   // This widget is the root of our application.
+
   @override
   Widget build(BuildContext context) {
     return const MaterialApp(
@@ -103,85 +103,5 @@ class MSFatigue extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       home: CheckCredentials(),
     );
-  }
-}
-
-/// A widget that checks for stored credentials and routes accordingly.
-class CheckCredentials extends StatefulWidget {
-  const CheckCredentials({super.key});
-
-  @override
-  State<CheckCredentials> createState() => _CheckCredentialsState();
-}
-
-class _CheckCredentialsState extends State<CheckCredentials> {
-  bool? credentialsExist;
-
-  Future<void> checkCredentials() async {
-    // Use SharedPreferences to read stored credentials.
-
-    final prefs = await SharedPreferences.getInstance();
-    final username = prefs.getString('msfatigue_username');
-    final password = prefs.getString('msfatigue_password');
-    final preferredName = prefs.getString('msfatigue_preferredName');
-
-    setState(() {
-      credentialsExist =
-          (username != null && password != null && preferredName != null);
-    });
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    checkCredentials();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // While checking, show a progress indicator.
-
-    if (credentialsExist == null) {
-      return const Scaffold(
-        body: Center(child: CircularProgressIndicator()),
-      );
-    }
-
-    // If credentials are missing, pop up a dialog then navigate to the CredentialsPage.
-
-    else if (credentialsExist == false) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        showDialog(
-          context: context,
-          barrierDismissible: false,
-          builder: (context) => AlertDialog(
-            title: const Text("Credentials Required"),
-            content: const Text(
-                "Please enter your username, password, and preferred name."),
-            actions: [
-              TextButton(
-                onPressed: () {
-                  Navigator.of(context).pop();
-                  Navigator.of(context).push(
-                    MaterialPageRoute(
-                      builder: (_) => const CredentialsPage(),
-                    ),
-                  );
-                },
-                child: const Text("OK"),
-              ),
-            ],
-          ),
-        );
-      });
-      return const Scaffold(
-        body: Center(child: Text("Please enter your credentials.")),
-      );
-    }
-    // If credentials exist, navigate to the main WelcomeScreen.
-
-    else {
-      return const WelcomeScreen();
-    }
   }
 }
